requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
setuptools~=54.2.0
mariadb
ipython
pandas
numpy
inflection
